Fix ignored engine/DateTimeExtensionTest
