+++ android/eir/src/main/java/org/smartregister/fhircore/eir/data/PatientRepository.kt
@@ -28,12 +28,12 @@ import org.smartregister.fhircore.eir.data.model.PatientItem
 import org.smartregister.fhircore.eir.ui.patient.details.AdverseEventItem
 import org.smartregister.fhircore.eir.ui.patient.register.PatientItemMapper
 import org.smartregister.fhircore.engine.data.domain.util.RegisterRepository
-import org.smartregister.fhircore.engine.util.DateUtils.toHumanDisplay
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 import org.smartregister.fhircore.engine.util.extension.countActivePatients
 import org.smartregister.fhircore.engine.util.extension.extractId
 import org.smartregister.fhircore.engine.util.extension.loadResource
 import org.smartregister.fhircore.engine.util.extension.searchActivePatients
+import org.smartregister.fhircore.engine.util.extension.toHumanDisplay
 
 class PatientRepository
 @Inject
