+++ android/engine/src/main/java/org/smartregister/fhircore/engine/util/DateUtils.kt
@@ -49,9 +49,6 @@ object DateUtils {
     return copy.before(DateTimeType.now())
   }
 
-  fun Date.toHumanDisplay(): String =
-    SimpleDateFormat("MMM d, yyyy h:mm:ss a", Locale.getDefault()).format(this)
-
   fun simpleDateFormat(pattern: String = "hh:mm aa, MMM d") =
     SimpleDateFormat(pattern, Locale.getDefault())
   fun Date?.makeItReadable(): String = if (this != null) simpleDateFormat.format(this) else "N/A"
